# less prints for better performance in output mode